+++ Assets/NeedsVsWants/Scripts/BillingSystem/CalendarEvent/SSS.cs
@@ -9,14 +9,14 @@ namespace NeedsVsWants.BillingSystem
     public class SSS : BillEvent
     {
         [SerializeField]
-        float _Amount;
+        double _Amount;
 
         public override bool IsWithinDate(DateTime dateTime)
         {
             return DateTime.DaysInMonth(dateTime.Year, dateTime.Month) == dateTime.Day;
         }
 
-        public override float CalculateBill(DateTime dateTime)
+        public override double CalculateBill(DateTime dateTime)
         {
             return _Amount;
         }
